+++ src/js/components/ChatView.js
@@ -5,15 +5,13 @@ import Message from './Message.js';
 import {activeChat, chats, lastSeenTimeChanged, setLatestSeen, setLatestCheckmark} from '../Chat.js';
 import Helpers from '../Helpers.js';
 import Notifications from '../Notifications.js';
+import Session from '../Session.js';
+
+const notificationServiceUrl = 'https://iris-notifications.herokuapp.com/notify';
 
 const submitButton = html`
   <button type="submit">
     <svg class="svg-inline--fa fa-w-16" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 0 486.736 486.736" style="enable-background:new 0 0 486.736 486.736;" xml:space="preserve" width="100px" height="100px" fill="#000000" stroke="#000000" stroke-width="0"><path fill="currentColor" d="M481.883,61.238l-474.3,171.4c-8.8,3.2-10.3,15-2.6,20.2l70.9,48.4l321.8-169.7l-272.4,203.4v82.4c0,5.6,6.3,9,11,5.9 l60-39.8l59.1,40.3c5.4,3.7,12.8,2.1,16.3-3.5l214.5-353.7C487.983,63.638,485.083,60.038,481.883,61.238z"></path></svg>
-  </button>`;
-
-const attachFileButton = html`
-  <button type="button" id="attach-file">
-    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path fill="currentColor" d="M21.586 10.461l-10.05 10.075c-1.95 1.949-5.122 1.949-7.071 0s-1.95-5.122 0-7.072l10.628-10.585c1.17-1.17 3.073-1.17 4.243 0 1.169 1.17 1.17 3.072 0 4.242l-8.507 8.464c-.39.39-1.024.39-1.414 0s-.39-1.024 0-1.414l7.093-7.05-1.415-1.414-7.093 7.049c-1.172 1.172-1.171 3.073 0 4.244s3.071 1.171 4.242 0l8.507-8.464c.977-.977 1.464-2.256 1.464-3.536 0-2.769-2.246-4.999-5-4.999-1.28 0-2.559.488-3.536 1.465l-10.627 10.583c-1.366 1.368-2.05 3.159-2.05 4.951 0 3.863 3.13 7 7 7 1.792 0 3.583-.684 4.95-2.05l10.05-10.075-1.414-1.414z"/></svg>
   </button>`;
 
 const emojiPickerButton = html`
@@ -30,6 +28,7 @@ class ChatView extends Component {
 
   componentDidMount() {
     localState.get('activeChat').on(activeChatId => {
+      Helpers.closeAttachmentsGallery();
       this.setState({});
       if (activeChatId && !subscribedToMsgs[activeChatId]) {
         const iv = setInterval(() => {
@@ -42,6 +41,98 @@ class ChatView extends Component {
         subscribedToMsgs[activeChatId] = true;
       }
     });
+    $(document).off('keyup').on('keyup', e => {
+      if (e.key === "Escape") { // escape key maps to keycode `27`
+        if ($('#attachment-preview.gallery:visible').length) {
+          Helpers.closeAttachmentsGallery();
+        } else {
+          this.closeAttachmentsPreview();
+        }
+      }
+    });
+  }
+
+  async onMsgFormSubmit(event) {
+    const pub = activeChat;
+    event.preventDefault();
+    chats[pub].msgDraft = null;
+    var text = $('#new-msg').val();
+    if (!text.length && !chats[pub].attachments) { return; }
+    chats[pub].setTyping(false);
+    var msg = {text};
+    if (chats[pub].attachments) {
+      msg.attachments = chats[pub].attachments;
+    }
+    const myKey = Session.getKey();
+    const shouldWebPush = (pub === myKey.pub) || !(chats[pub].online && chats[pub].online.isOnline);
+    if (shouldWebPush && chats[pub].webPushSubscriptions) {
+      const subscriptions = [];
+      const participants = Object.keys(chats[pub].webPushSubscriptions);
+      for (let i = 0; i < participants.length; i++) {
+        const participant = participants[i];
+        const secret = await chats[pub].getSecret(participant);
+        const myName = Session.getMyName();
+        const titleText = chats[pub].uuid ? chats[pub].name : myName;
+        const bodyText = chats[pub].uuid ? `${myName}: ${text}` : text;
+        const payload = {
+          title: await Gun.SEA.encrypt(titleText, secret),
+          body: await Gun.SEA.encrypt(bodyText, secret),
+          from:{pub: myKey.pub, epub: myKey.epub}
+        };
+        chats[pub].webPushSubscriptions[participant].forEach(s => subscriptions.push({subscription: s, payload}));
+      }
+      fetch(notificationServiceUrl, {
+        method: 'POST',
+        body: JSON.stringify({subscriptions}),
+        headers: {
+          'content-type': 'application/json'
+        }
+      });
+    }
+    chats[pub].send(msg);
+    this.closeAttachmentsPreview();
+    $('#new-msg').val('');
+  }
+
+  attachFileClicked(event) {
+    console.log('hi');
+    event.preventDefault();
+    $('#attachment-input').click();
+  }
+
+  openAttachmentsPreview() {
+    console.log('hohoo');
+    $('#floating-day-separator').remove();
+    var attachmentsPreview = $('#attachment-preview');
+    attachmentsPreview.removeClass('gallery');
+    attachmentsPreview.empty();
+    var closeBtn = $('<button>').text(t('cancel')).click(this.closeAttachmentsPreview);
+    attachmentsPreview.append(closeBtn);
+
+    var files = $('#attachment-input')[0].files;
+    if (files) {
+      attachmentsPreview.show();
+      $('#message-list').hide();
+      for (var i = 0;i < files.length;i++) {
+        Helpers.getBase64(files[i]).then(base64 => {
+          chats[activeChat].attachments = chats[activeChat].attachments || [];
+          chats[activeChat].attachments.push({type: 'image', data: base64});
+          var preview = Helpers.setImgSrc($('<img>'), base64);
+          attachmentsPreview.append(preview);
+        });
+      }
+      $('#new-msg').focus();
+    }
+  }
+
+  closeAttachmentsPreview() {
+    $('#attachment-preview').hide();
+    $('#attachment-preview').removeClass('gallery');
+    $('#message-list').show();
+    if (activeChat) {
+      chats[activeChat].attachments = null;
+    }
+    Helpers.scrollToMessageListBottom();
   }
 
   subscribeToMsgs(pub) {
@@ -130,9 +221,11 @@ class ChatView extends Component {
         <p><button class="copy-chat-link">${t('copy_your_chat_link')}</button></p>
       </div>
       <div class="message-form" style="display:none">
-        <form autocomplete="off">
-          <input name="attachment-input" type="file" class="hidden" id="attachment-input" accept="image/*" multiple/>
-          ${attachFileButton}
+        <form autocomplete="off" onSubmit=${e => this.onMsgFormSubmit(e)}>
+          <input name="attachment-input" type="file" class="hidden" id="attachment-input" accept="image/*" multiple onChange=${() => this.openAttachmentsPreview()}/>
+          <button type="button" id="attach-file" onClick=${this.attachFileClicked}>
+            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path fill="currentColor" d="M21.586 10.461l-10.05 10.075c-1.95 1.949-5.122 1.949-7.071 0s-1.95-5.122 0-7.072l10.628-10.585c1.17-1.17 3.073-1.17 4.243 0 1.169 1.17 1.17 3.072 0 4.242l-8.507 8.464c-.39.39-1.024.39-1.414 0s-.39-1.024 0-1.414l7.093-7.05-1.415-1.414-7.093 7.049c-1.172 1.172-1.171 3.073 0 4.244s3.071 1.171 4.242 0l8.507-8.464c.977-.977 1.464-2.256 1.464-3.536 0-2.769-2.246-4.999-5-4.999-1.28 0-2.559.488-3.536 1.465l-10.627 10.583c-1.366 1.368-2.05 3.159-2.05 4.951 0 3.863 3.13 7 7 7 1.792 0 3.583-.684 4.95-2.05l10.05-10.075-1.414-1.414z"/></svg>
+          </button>
           ${emojiPickerButton}
           <input id="new-msg" type="text" placeholder="${t('type_a_message')}" autocomplete="off" autocorrect="off" autocapitalize="sentences" spellcheck="off"/>
           ${submitButton}
