+++ src/js/components/Message.js
@@ -1,15 +1,56 @@
 import { html, Component } from '../lib/htm.preact.js';
 import Helpers from '../Helpers.js';
-import Gallery from '../Gallery.js';
 import {chats} from '../Chat.js';
 
 const autolinker = new Autolinker({ stripPrefix: false, stripTrailingSlash: false});
+const ANIMATE_DURATION = 200;
+
 
 const seenIndicator = html`<span class="seen-indicator"><svg version="1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 59 42"><polygon fill="currentColor" points="40.6,12.1 17,35.7 7.4,26.1 4.6,29 17,41.3 43.4,14.9"></polygon><polygon class="iris-delivered-checkmark" fill="currentColor" points="55.6,12.1 32,35.7 29.4,33.1 26.6,36 32,41.3 58.4,14.9"></polygon></svg></span>`;
 
 function addMention(name) {
   $('#new-msg').val($('#new-msg').val().trim() + ` @${name} `);
   $('#new-msg').focus();
+}
+
+function openAttachmentsGallery(msg, event) {
+  $('#floating-day-separator').remove();
+  var attachmentsPreview = $('#attachment-preview');
+  attachmentsPreview.addClass('gallery');
+  attachmentsPreview.empty();
+  attachmentsPreview.fadeIn(ANIMATE_DURATION);
+  var left, top, width, img;
+
+  if (msg.attachments) {
+    msg.attachments.forEach(a => {
+      if (a.type.indexOf('image') === 0 && a.data) {
+        img = Helpers.setImgSrc($('<img>'), a.data);
+        if (msg.attachments.length === 1) {
+          attachmentsPreview.css({'justify-content': 'center'});
+          var original = $(event.target);
+          left = original.offset().left;
+          top = original.offset().top - $(window).scrollTop();
+          width = original.width();
+          var transitionImg = img.clone().attr('id', 'transition-img').data('originalDimensions', {left,top,width});
+          transitionImg.css({position: 'fixed', left, top, width, 'max-width': 'none', 'max-height': 'none'});
+          img.css({visibility: 'hidden', 'align-self': 'center'});
+          attachmentsPreview.append(img);
+          $('body').append(transitionImg);
+          var o = img.offset();
+          transitionImg.animate({width: img.width(), left: o.left, top: o.top}, {duration: ANIMATE_DURATION, complete: () => {
+            img.css({visibility: 'visible'});
+            transitionImg.hide();
+          }});
+        } else {
+          attachmentsPreview.css({'justify-content': ''});
+          attachmentsPreview.append(img);
+        }
+      }
+    })
+  }
+  $('#attachment-preview').one('click', () => {
+    Helpers.closeAttachmentsGallery();
+  });
 }
 
 class Message extends Component {
@@ -45,7 +86,7 @@ class Message extends Component {
         <div class="msg-content">
           ${name && this.props.showName && html`<small onclick=${() => addMention(name)} class="msgSenderName" style="color: ${color}">${name}</small>`}
           ${this.props.attachments && this.props.attachments.map(a =>
-            html`<img src=${a.data} onclick=${e => { Gallery.openAttachmentsGallery(this.props, e); }}/>` // escape a.data
+            html`<img src=${a.data} onclick=${e => { openAttachmentsGallery(this.props, e); }}/>` // escape a.data
           )}
           <div class="text ${emojiOnly && 'emoji-only'}" dangerouslySetInnerHTML=${{ __html: innerHTML }}>
           </div>
